refactor: break circular dependency in usage modules
- Create usage/types.ts with shared FreeTierLimit and FreeTierUsage interfaces
- Update usage/limits.ts to import types from ./types.ts
- Update usage/formatters.ts to remove duplicate interface definition
- Resolves Madge circular dependency: limits.ts → formatters.ts → limits.ts

Console log review: No action needed - all console statements are:
- In CLI scripts (appropriate for user feedback)
- In logger implementation (necessary)
- None in production provider code
